requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
pandas
numpy
nltk
joblib
scikit-learn
pdfminer.six
